mouse pointer broadcast function moved
